@@ -4,4 +4,4 @@ const config = {
   },
 };
 
-module.exports = config;
+export default config;
